fixed http status codes, renamed file for consistency
