+++ src/api/models/v2/course.model.ts
@@ -1,5 +1,5 @@
 import { DataTypes, Model, Optional } from 'sequelize';
-import { sequelize } from '../../../config/databasev2';
+import { sequelize } from '../../../config/database_v2';
 
 import { Organization } from './organization.model';
