+++ src/api/models/v2/module.model.ts
@@ -1,5 +1,5 @@
 import { DataTypes, Model, Optional } from 'sequelize';
-import { sequelize } from '../../../config/databasev2';
+import { sequelize } from '../../../config/database_v2';
 import { Course } from './course.model';
 
 export interface ModuleAttributes {
